client: Put client-api specific functionality behind a feature flag
